+++ src/star_teller/resources/icons/star.png
