+++ src/star_teller/ui/star_teller.ui
